Test app code rearranged. AO random vector scale related bug fixed. AO image y axis flipped. Depth shader made to also record normal vectors for all fragments. Types added for float limits. Random class added for stochastic coding. Added packed texture type for render targets. Frame number now starts at 0. Code for layout transition of render targets moved to vulkan texture. AO render view completed. Is render target added as configurable property for textures. Stack allocator bugs fixed. All MemoryTag::Temp deallocations made to be preformed in reverse order. Generation of SSAO noise texture is now random. Render pass re-initialization no longer crashes the engine. Shader descriptor set can now deal with packed textures. Double free on texture deletion fixed. Writable texture creation bug where mip-mapping was always true fixed.

+++ include/renderer/vulkan/vulkan_types.hpp
@@ -73,7 +73,7 @@ struct PhysicalDeviceInfo {
  */
 struct VulkanCommandBuffer {
     /// @brief Currently recorded frame
-    uint32                   current_frame;
+    uint8                    current_frame;
     /// @brief Handle to the currently recorded buffer
     const vk::CommandBuffer* handle;
 
@@ -84,7 +84,7 @@ struct VulkanCommandBuffer {
      * @brief Flushes contents and resets recording
      * @param current_frame Index of the next frame
      */
-    void reset(const uint32 current_frame);
+    void reset(const uint8 current_frame);
 
   private:
     const Vector<vk::CommandBuffer> _buffers;
